requirements.txt
streamlit==0.60.0
pandas==1.0.3
joblib==0.14.1
Pillow==7.1.2

matplotlib==3.3.2
torchnet==0.0.4
